+++ TransIT.API.sln
@@ -3,7 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.489
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TransIT.API", "TransIT.API\TransIT.API.csproj", "{73CF8B65-616E-4999-8A85-C1ECA7F644F8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransIT.DAL", "TransIT.DAL\TransIT.DAL.csproj", "{E524D22E-7F05-4AFC-9082-C87C8E3CA9DB}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransIT.API", "TransIT.API\TransIT.API.csproj", "{BEFED30C-E61F-48AE-BF4C-3751ED293358}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TransIT.BLL", "TransIT.BLL\TransIT.BLL.csproj", "{3BB6FAF4-8D54-4428-AD1F-923A14D3EEE1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +15,18 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{73CF8B65-616E-4999-8A85-C1ECA7F644F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{73CF8B65-616E-4999-8A85-C1ECA7F644F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{73CF8B65-616E-4999-8A85-C1ECA7F644F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{73CF8B65-616E-4999-8A85-C1ECA7F644F8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E524D22E-7F05-4AFC-9082-C87C8E3CA9DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E524D22E-7F05-4AFC-9082-C87C8E3CA9DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E524D22E-7F05-4AFC-9082-C87C8E3CA9DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E524D22E-7F05-4AFC-9082-C87C8E3CA9DB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BEFED30C-E61F-48AE-BF4C-3751ED293358}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BEFED30C-E61F-48AE-BF4C-3751ED293358}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BEFED30C-E61F-48AE-BF4C-3751ED293358}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BEFED30C-E61F-48AE-BF4C-3751ED293358}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3BB6FAF4-8D54-4428-AD1F-923A14D3EEE1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3BB6FAF4-8D54-4428-AD1F-923A14D3EEE1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3BB6FAF4-8D54-4428-AD1F-923A14D3EEE1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3BB6FAF4-8D54-4428-AD1F-923A14D3EEE1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
