+++ TransIT.BLL/TransIT.BLL.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\TransIT.DAL\TransIT.DAL.csproj" />
+  </ItemGroup>
+
+</Project>
