+++ TransIT.DAL/Models/Malfunction.cs
@@ -1,22 +1,18 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace TransIT.Models.Entities
+namespace TransIT.DAL.Models
 {
-    public partial class Vehicle
+    public partial class Malfunction
     {
-        public Vehicle()
+        public Malfunction()
         {
             Issue = new HashSet<Issue>();
         }
 
         public int Id { get; set; }
-        public int VehicleTypeId { get; set; }
-        public string Vincode { get; set; }
-        public string InventoryId { get; set; }
-        public string RegNum { get; set; }
-        public string Brand { get; set; }
-        public string Model { get; set; }
+        public string Name { get; set; }
+        public int MalfunctionSubgroupId { get; set; }
         public DateTime? CreateDate { get; set; }
         public DateTime? ModDate { get; set; }
         public int CreateId { get; set; }
@@ -24,7 +20,6 @@ namespace TransIT.Models.Entities
 
         public virtual User Create { get; set; }
         public virtual User Mod { get; set; }
-        public virtual VehicleType VehicleType { get; set; }
         public virtual ICollection<Issue> Issue { get; set; }
     }
 }
