+++ TransIT.DAL/Models/MalfunctionGroup.cs
@@ -1,10 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace TransIT.Models.Entities
+namespace TransIT.DAL.Models
 {
-    public partial class Role
+    public partial class MalfunctionGroup
     {
+        public MalfunctionGroup()
+        {
+            MalfunctionSubgroup = new HashSet<MalfunctionSubgroup>();
+        }
+
         public int Id { get; set; }
         public string Name { get; set; }
         public DateTime? CreateDate { get; set; }
@@ -12,7 +17,6 @@ namespace TransIT.Models.Entities
         public int CreateId { get; set; }
         public int ModId { get; set; }
 
-        public virtual User Create { get; set; }
-        public virtual User Mod { get; set; }
+        public virtual ICollection<MalfunctionSubgroup> MalfunctionSubgroup { get; set; }
     }
 }
