--- TransIT.Models/Entities/MalfunctionGroup.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace TransIT.Models.Entities
-{
-    public partial class MalfunctionGroup
-    {
-        public MalfunctionGroup()
-        {
-            MalfunctionSubgroup = new HashSet<MalfunctionSubgroup>();
-        }
-
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public DateTime? CreateDate { get; set; }
-        public DateTime? ModDate { get; set; }
-        public int CreateId { get; set; }
-        public int ModId { get; set; }
-
-        public virtual ICollection<MalfunctionSubgroup> MalfunctionSubgroup { get; set; }
-    }
-}
